+++ addOns/requester/CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-
+### Fixed
+- Save Requester panel and Manual Request Editor dialog options (Issue 6985).
 
 ## [7.9.0] - 2025-12-15
 ### Changed